Updated deliverable lists with new columns
