+++ src/scenes/AcceleratorRouter/Deliverables/DeliverablesList.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo, useRef, useState } from 'react';
 
 import { Grid, Typography } from '@mui/material';
-import { Separator, TableColumnType } from '@terraware/web-components';
+import { Separator } from '@terraware/web-components';
 
 import DeliverablesTable from 'src/components/DeliverablesTable';
 import PageHeader from 'src/components/PageHeader';
@@ -14,47 +14,6 @@ import AcceleratorMain from 'src/scenes/AcceleratorRouter/AcceleratorMain';
 import strings from 'src/strings';
 import theme from 'src/theme';
 import { SearchNodePayload } from 'src/types/Search';
-
-const columns = (activeLocale: string | null): TableColumnType[] =>
-  activeLocale
-    ? [
-        {
-          key: 'name',
-          name: strings.DELIVERABLE_NAME,
-          type: 'string',
-        },
-        {
-          key: 'dueDate',
-          name: strings.DUE_DATE,
-          type: 'date',
-        },
-        {
-          key: 'type',
-          name: strings.TYPE,
-          type: 'string',
-        },
-        {
-          key: 'numDocuments',
-          name: strings.DOCUMENTS,
-          type: 'number',
-        },
-        {
-          key: 'category',
-          name: strings.CATEGORY,
-          type: 'string',
-        },
-        {
-          key: 'projectName',
-          name: strings.PROJECT,
-          type: 'string',
-        },
-        {
-          key: 'status',
-          name: strings.STATUS,
-          type: 'string',
-        },
-      ]
-    : [];
 
 const DeliverablesList = () => {
   const { activeLocale } = useLocalization();
@@ -118,7 +77,6 @@ const DeliverablesList = () => {
       {/* -1 for "non-organization scoped search" IE admin search */}
       <DeliverablesTable
         acceleratorProjects={acceleratorProjects}
-        columns={columns}
         extraTableFilters={extraTableFilters}
         isAcceleratorRoute={true}
         organizationId={-1}
